+++ project/src/store/favorites/favourites-data.ts
@@ -1,11 +1,12 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { NameSpace } from '../../const';
+import { FetchStatus, NameSpace } from '../../const';
 import { FavouriteData } from '../../types/store';
 import { fetchFavouriteOffers } from '../api-actions';
 
 const initialState: FavouriteData = {
   favouriteOffers: [],
   isFavouriteOffersLoaded: false,
+  status: FetchStatus.Idle
 };
 
 export const favouritesData = createSlice({
@@ -16,13 +17,16 @@ export const favouritesData = createSlice({
     builder
       .addCase(fetchFavouriteOffers.pending, (state) => {
         state.isFavouriteOffersLoaded = false;
+        state.status = FetchStatus.Pending;
       })
       .addCase(fetchFavouriteOffers.rejected, (state) => {
         state.isFavouriteOffersLoaded = true;
+        state.status = FetchStatus.Rejected;
       })
       .addCase(fetchFavouriteOffers.fulfilled, (state, action) => {
         state.isFavouriteOffersLoaded = true;
         state.favouriteOffers = action.payload;
+        state.status = FetchStatus.Success;
       });
   }
 });
